+++ d60.EventSorcerer/Numbers/ISequenceNumberGenerator.cs
@@ -3,10 +3,11 @@
 namespace d60.EventSorcerer.Numbers
 {
     /// <summary>
-    /// Implement this to provide logic on how stuff is sequentially numbered
+    /// Implement this to provide logic on how stuff is sequentially numbered. Sequence number generators must ALWAYS start with the number 0 for each
+    /// unique aggregate root ID
     /// </summary>
     public interface ISequenceNumberGenerator
     {
-        int Next(Guid aggregateRootId);
+        long Next(Guid aggregateRootId);
     }
 }
